==> preprocessing/03_basefile.Rmd <==
---
title: "OOPA Risk of Material Breach Basefile"
author: "Nathaniel Olin"
date: "`r date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

This script creates a BRT-level basefile from the cleaned `parcels.csv` and `OOPA_report_clean.csv` datasets.

```{r, include = F}
options(cli.width = 100)
```


```{r}
library(tidyverse)
```

# Load data

Load full list of BRTs and address data.

```{r}
tips <- read_csv(file.path('..', 'data', 'parcels.csv'))
# Convert BRT to integer to drop leading zeroes, for matching
table(nchar(tips$brt))
tips$brt <- as.integer(tips$brt)
table(nchar(tips$brt))
```

Load OOPA case-level data

```{r}
oopa <- read_csv(file.path('..', 'data', 'OOPA_report_clean.csv'))
```

Load census data

## Census data

```{r}
census <- read_csv(file.path('..', 'data', 'census_tract_data.csv'))
table(nchar(census$brt))
census$brt <- as.integer(census$brt)
table(nchar(census$brt))
```

## Data quality checks

BRT number should be unique

```{r}
stopifnot(!duplicated(tips$brt))
stopifnot(!duplicated(census$brt))
```

Case should be unique on OOPA file

```{r}
stopifnot(!duplicated(oopa$case))
```

## Merge

```{r}
c('OPA rows' = nrow(tips),
  'OPA unique' = length(unique(tips$brt)),
  'OOPA rows' = nrow(oopa),
  'OOPA unique' = length(unique(oopa$brt)),
  'Census rows' = nrow(census),
  'Census unique' = length(unique(census$brt)))

dat <- oopa %>% 
  full_join(tips, by = 'brt') %>%
  left_join(census, by = 'brt')

c('Final rows' = nrow(dat),
  'Final unique BRT' = length(unique(dat$brt)))
```

```{r}
c('Final dataset' = nrow(dat),
  'In both datasets' = sum(dat$brt %in% oopa$brt & dat$brt %in% tips$brt),
  'In OOPA not OPA (0?)' = sum(dat$brt %in% oopa$brt & ! dat$brt %in% tips$brt),
  'In OPA not OOPA (big)' = sum(dat$brt %in% tips$brt & ! dat$brt %in% oopa$brt),
  'In Census not OPA' = sum(dat$brt %in% census$brt & ! dat$brt %in% tips$brt),
  'In OPA not Census' = sum(dat$brt %in% tips$brt & ! dat$brt %in% census$brt))

stopifnot(tips$brt %in% dat$brt,
          oopa$brt %in% dat$brt)
```

```{r}
rm(oopa, tips, census)
```

# Collapse to BRT level

How many BRTs appear more than once?

```{r}
dat %>%
  filter(duplicated(brt)) %>%
  select(brt) %>% unique() %>% nrow()
```

If there are multiple cases with the same BRT:

* If only one is ongoing, take that.
* If none are ongoing, take the most recent one.
* If multiple are ongoing, check the data / report - shouldn't be possible.

```{r}
dat <- dat %>%
  mutate(active = status == '520 Ongoing') %>%
  arrange(brt, desc(active), desc(agree_start)) %>%
  group_by(brt) %>%
  mutate(brt_rank = 1:n()) %>%
  ungroup()

dat %>% 
  filter(brt %in% dat$brt[duplicated(dat$brt)]) %>%
  select(brt, agree_start, brt_rank)
```

```{r}
reject <- subset(dat, !is.na(brt) & brt_rank > 1)

# Did any active OOPAs get kicked out when de-duplicating by BRT?
table(reject$status)
dat[dat$brt %in% reject$brt[reject$status == '520 Ongoing'], ]
# Not enough to matter
```

Deduplicate

```{r}
dat <- subset(dat, is.na(brt) | brt_rank == 1)

stopifnot(!duplicated(dat$brt))
dat$brt_rank <- NULL
dat$active <- NULL
```

# New variables

## Current OOPA status
    
* Not in OOPA
* Currently in OOPA, good standing
* Currently in OOPA, in material breach
* Previously in OOPA, completed
* Previously in OOPA, defaulted
    
```{r}
dat$status[is.na(dat$status)] <- 'Not in OOPA'

# Did people fall behind on payments?
dat$last_date <- as.Date(NA)
dat$last_date[!is.na(dat$default_dt)] <- dat$default_dt[!is.na(dat$default_dt)]
dat$last_date[!is.na(dat$complete_dt)] <- dat$complete_dt[!is.na(dat$complete_dt)]
dat$last_date[is.na(dat$last_date) & !is.na(dat$case)] <- Sys.Date()

dat <- dat %>%
  mutate(months_in_agree = 
           (lubridate::year(last_date) * 12 + lubridate::month(last_date)) -
           (lubridate::year(agree_start) * 12 + lubridate::month(agree_start)),
         num_diff = months_in_agree - num_pmts)

# dat$status[dat$status == '520 Ongoing' & 
#              dat$num_diff > 1] <- '520 Ongoing, Breached'
```

## Bin census variables

```{r}
# Binning function that mean-imputes missing values
bin <- function(x, n = 4, missing.fun = function(x) mean(x, na.rm = T)){
  x.na <- missing.fun(x)
  cat("Imputing", scales::comma(sum(is.na(x))), "values as", x.na, "\n")
  x[is.na(x)] <- x.na
  cut(x, breaks = n)
}
```

```{r}
dat <- dat %>%
  mutate(bin_no_hs = bin(pct_no_hs),
         bin_college = bin(pct_college),
         bin_non_english = bin(pct_non_english))
```

Add in missing values in stratification variables (or drop when small)

```{r}
dat$tier[is.na(dat$tier)] <- 'Missing'

table(missing_mail = is.na(dat$different_mail), 
      in_oopa = is.na(dat$case))
# Drop these records
dat <- subset(dat, !is.na(different_mail))

table(missing_zip = is.na(dat$property_zip5),
      in_oopa = is.na(dat$case))
# Don't want to drop active OOPA cases so
dat$property_zip5[is.na(dat$property_zip5)] <- 'Missing'
```

# Write out

Check overall data and data with OOPA numbers

```{r}
dat %>% print(n = 4, width = Inf)
dat %>% filter(!is.na(case)) %>% print(n = 4, width = Inf)
write_csv(dat, file.path('..', 'data', 'basefile.csv'))
```
